Unify Email entry-point tests across Create/TryCreate/ToEmail/AsEmail
The four entry points (Email.Create, Email.TryCreate, string.ToEmail,
string.AsEmail) share one validation contract — collapse the per-entry
tests into one Theory per case bucket (invalid / max-length boundary /
valid) so the four wrappers stay lock-stepped. MailAddress.Create and
MailAddress.TryCreate keep their own tests since they are BCL shims with
deliberately different semantics (FormatException, no MaxLength check).
